Switch player after each move.
