+++ src/main/java/gui_app_runner.java
@@ -1,6 +1,3 @@
-import game.Board;
-import game.Game;
-import game.Mark;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
@@ -18,11 +15,7 @@ public class gui_app_runner extends Application {
     @Override
     public void start(Stage primaryStage) throws IOException {
         primaryStage.setTitle("Tic Tac Toe");
-        Parent root = FXMLLoader.load(getClass().getResource("Display.fxml"));
-        int board = new Board().size;
-//        Mark currentMark = new Game(new Display(), new Board()).currentPlayer.getMark();
-//        Display display = new Display(board, currentMark);
-//        primaryStage.setScene(display);
+        Parent root = FXMLLoader.load(getClass().getResource("Controller.fxml"));
         primaryStage.setScene(new Scene(root, 300, 275));
         primaryStage.show();
     }
