+++ src/main/resources/Controller.fxml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<AnchorPane xmlns="http://javafx.com/javafx"
+            xmlns:fx="http://javafx.com/fxml"
+            fx:controller="Controller"
+            prefHeight="400.0" prefWidth="600.0">
+
+    <GridPane xmlns:fx="http://javafx.com/fxml" alignment="center" hgap="5" vgap="5">
+        <Button text="1" onAction="#makeMove" alignment="BASELINE_LEFT" GridPane.rowIndex="1" GridPane.columnIndex="1" fx:id="one"/>
+        <Button text="2" onAction="#makeMove" alignment="BASELINE_LEFT" GridPane.rowIndex="1" GridPane.columnIndex="2" fx:id="two"/>
+        <Button text="3" onAction="#makeMove" alignment="BASELINE_LEFT" GridPane.rowIndex="1" GridPane.columnIndex="3" fx:id="three"/>
+        <Button text="4" onAction="#makeMove" alignment="BASELINE_LEFT" GridPane.rowIndex="2" GridPane.columnIndex="1" fx:id="four"/>
+        <Button text="5" onAction="#makeMove" alignment="BASELINE_LEFT" GridPane.rowIndex="2" GridPane.columnIndex="2" fx:id="five"/>
+        <Button text="6" onAction="#makeMove" alignment="BASELINE_LEFT" GridPane.rowIndex="2" GridPane.columnIndex="3" fx:id="six"/>
+        <Button text="7" onAction="#makeMove" alignment="BASELINE_LEFT" GridPane.rowIndex="3" GridPane.columnIndex="1" fx:id="seven"/>
+        <Button text="8" onAction="#makeMove" alignment="BASELINE_LEFT" GridPane.rowIndex="3" GridPane.columnIndex="2" fx:id="eight"/>
+        <Button text="9" onAction="#makeMove" alignment="BASELINE_LEFT" GridPane.rowIndex="3" GridPane.columnIndex="3" fx:id="nine"/>
+    </GridPane>
+</AnchorPane>
